feat: notification 경로 수정
